enforce guest user

+++ server/routes/medias-routes.js
@@ -188,30 +188,32 @@ let downloadFile = (media) => {
 
 
 router.post('/', authenticate, upload, (req, res) => {
-  let body = _.pick(req.passedMedia, mediaInsertFields);
-  console.log('body', body);
-  let media = new Media(body);
-  console.log('body.photoInfo', body.photoInfo);
-  if (body.photoInfo && body.photoInfo.location && body.photoInfo.isUrl) {
-    media.location = body.photoInfo.location;
-    media.isUrl = true;
-    media.mimeType = body.photoInfo.mimeType;
+  if (!req.loggedInUser.guestUser) {
+    let body = _.pick(req.passedMedia, mediaInsertFields);
+    console.log('body', body);
+    let media = new Media(body);
+    console.log('body.photoInfo', body.photoInfo);
+    if (body.photoInfo && body.photoInfo.location && body.photoInfo.isUrl) {
+      media.location = body.photoInfo.location;
+      media.isUrl = true;
+      media.mimeType = body.photoInfo.mimeType;
+    };
+
+    media._creator = req.loggedInUser._creatorRef;
+    media.isProfilePic = false;
+    media.addedDate = new Date().getTime();
+    media.comments = [];
+    media._id = new ObjectID();
+    console.log('media', media);
+
+    media.save().then((newMedia) => {
+      console.log('media2', newMedia);
+      res.send(_.pick(newMedia, mediaOutFields));
+    }, (e) => {
+      console.log('media.save() e', e);
+      res.status(400).send();
+    });
   };
-
-  media._creator = req.loggedInUser._creatorRef;
-  media.isProfilePic = false;
-  media.addedDate = new Date().getTime();
-  media.comments = [];
-  media._id = new ObjectID();
-  console.log('media', media);
-
-  media.save().then((newMedia) => {
-    console.log('media2', newMedia);
-    res.send(_.pick(newMedia, mediaOutFields));
-  }, (e) => {
-    console.log('media.save() e', e);
-    res.status(400).send();
-  });
 });
 
 
@@ -323,40 +325,42 @@ router.get('/:id', authenticate, (req, res) => {
 });
 
 router.delete('/:id', authenticate, (req, res) => {
-  let {
-    id
-  } = req.params;
+  if (!req.loggedInUser.guestUser) {
+    let {
+      id
+    } = req.params;
 
 
-  if (!ObjectID.isValid(id)) {
-    return res.status(404).send({
-      error: "Media ID is invalid"
-    });
-  };
-
-  let medias = {
-    '_id': id
-  };
-  if (!req.loggedInUser.adminUser) {
-    medias._creator = req.loggedInUser._creatorRef;
-  }
-
-  Media.findOneAndRemove(medias).then((media) => {
-
-    if (media) {
-      googleCloudApi.deleteFile(media.location);
-      res.send({
-        media
+    if (!ObjectID.isValid(id)) {
+      return res.status(404).send({
+        error: "Media ID is invalid"
       });
-    } else {
-      res.status(404).send({
-        error: "media not found for id"
-      });
+    };
+
+    let medias = {
+      '_id': id
+    };
+    if (!req.loggedInUser.adminUser) {
+      medias._creator = req.loggedInUser._creatorRef;
     }
 
-  }, (e) => {
-    res.status(400).send(e);
-  });
+    Media.findOneAndRemove(medias).then((media) => {
+
+      if (media) {
+        googleCloudApi.deleteFile(media.location);
+        res.send({
+          media
+        });
+      } else {
+        res.status(404).send({
+          error: "media not found for id"
+        });
+      }
+
+    }, (e) => {
+      res.status(400).send(e);
+    });
+  };
 });
 
 
@@ -445,46 +449,48 @@ let updateMedias = (res, body, medias, commentId) => {
 
 router.patch('/:id', authenticate, upload, (req, res) => {
   console.log("router.patch1", req.passedMedia);
-  let {
-    id
-  } = req.params;
+  if (!req.loggedInUser.guestUser) {
+    let {
+      id
+    } = req.params;
 
-  let body = _.pick(req.passedMedia, mediaUpdateFields);
+    let body = _.pick(req.passedMedia, mediaUpdateFields);
 
-  console.log("router.patch2", body);
+    console.log("router.patch2", body);
 
-  if (!ObjectID.isValid(id)) {
-    return res.status(404).send({
-      error: "Media ID is invalid"
-    });
-  };
-
-  let medias = {
-    '_id': id
-  };
-
-
-  if (body.comment) {
-
-    let mediaCommentObj = {
-      'comment': body.comment
-    };
-    let comment = new Comment(mediaCommentObj);
-
-    comment._creator = req.loggedInUser._creatorRef;
-    comment.commentDate = new Date().getTime();
-    console.log('comment', comment);
-
-    comment.save().then((comment) => {
-      updateMedias(res, body, medias, comment._id);
-    });
-  } else {
-
-    if (!req.loggedInUser.adminUser) {
-      medias._creator = req.loggedInUser._creatorRef;
+    if (!ObjectID.isValid(id)) {
+      return res.status(404).send({
+        error: "Media ID is invalid"
+      });
     };
 
-    updateMedias(res, body, medias, null);
+    let medias = {
+      '_id': id
+    };
+
+
+    if (body.comment) {
+
+      let mediaCommentObj = {
+        'comment': body.comment
+      };
+      let comment = new Comment(mediaCommentObj);
+
+      comment._creator = req.loggedInUser._creatorRef;
+      comment.commentDate = new Date().getTime();
+      console.log('comment', comment);
+
+      comment.save().then((comment) => {
+        updateMedias(res, body, medias, comment._id);
+      });
+    } else {
+
+      if (!req.loggedInUser.adminUser) {
+        medias._creator = req.loggedInUser._creatorRef;
+      };
+
+      updateMedias(res, body, medias, null);
+    };
   };
 });
